# Improve `par-each` to output values in the same order as `each`