# Description

I noticed that the issue with `par-each`'s output ordering is because it uses `par_bridge()`. If it used more "native" Rayon methods, it'd preserve the order.

So, I changed it so that for each datatype that can't use `to_par_iter()` is collected to Vec first, so that it can. The performance impact, based on my tests, is below notice - for Ranges (`benchmark { 1..999 | par-each { $in } }`) it only increased from 4ms → 6ms. For Lists (`benchmark { 1..999 | each { $in } | par-each { $in / 2 } }`), performance improved from 41ms → 36ms (because it never needed `par_bridge()` in the first place). (However, I didn't test ListStreams or ExternalStreams because I don't know how exactly to do that.)

# User-Facing Changes

Ordering of `par-each` output is no longer arbitrary.

# Tests + Formatting

Don't forget to add tests that cover your changes.

Make sure you've run and fixed any issues with these commands:

- `cargo fmt --all -- --check` to check standard code formatting (`cargo fmt --all` applies these changes)
- `cargo clippy --workspace -- -D warnings -D clippy::unwrap_used -A clippy::needless_collect` to check that you're using the standard code style
- `cargo test --workspace` to check that all tests pass

# After Submitting

If your PR had any user-facing changes, update [the documentation](https://github.com/nushell/nushell.github.io) after the PR is merged, if necessary. This will help us keep the docs up to date.