@jntrnr 2022-11-26
Collecting into a vec should be something we avoid doing unless it's absolutely necessary. For this, `par-each` should work similarly to `each`, and have similar capabilities.

`each` can work over infinite streams, letting you fetch as many values as you want from the resulting stream. `par-each` should do the same.