@webbedspace 2022-11-26
Well, [there was already a .collect() *after* the parallel .map()](https://github.com/nushell/nushell/pull/7244/files#diff-8e27646f010a6e78106d9b558c8b434b34cfe92f2246dc3af7e4e56ccc67e071R166), so all I've done here is add a second one in front of the .map(). This PR doesn't actually change `par-each`'s infinite-stream-handling (in)ability on that front (although I admit that would be a desirable feature to add).