@jntrnr 2022-11-26
Now I'm trying to remember why we collect after the mapping. It may be to get around some type compatibility issues. If so, that's definitely unfortunate.