@jntrnr 2022-11-26
I asked on the rayon gitter to see if they know of a way to work with unbounded iterators. Hopefully there are some ways and we can move the technique we currently use to something that fits nushell a bit better.