@rgwood 2023-01-04
Sorry we've been so slow on this one.

I'm a little reluctant to merge this. In my mind, `par-each` threads should shove results out the door as soon as they're processed; the fact that we currently collect all results first is a bug. Even if `par-each` doesn't handle streaming well now, it should do so in the future.

As I understand it, this PR guarantees ordering of results but we'll have to take back that guarantee when we make `par-each` handle streaming properly. Feels like making promises we can't keep. Let me know if I've misunderstood the situation.