@webbedspace 2023-01-06
I've been thinking about the use-cases of `par-each`… I think there's still a large class of problems which depend on value ordering but would still need the parallelism, like "turn this list of URLs into their HTTP responses" or "turn this list of files into their first 2 lines", and which would need to recover the ordering after every operation is completed, thus not benefiting from `par-each` streaming the output. And moreover, those operations which *would* benefit from par-each streaming the output would be suited to simply moving the remaining operations into the `par-each` closure, like `par-each {|e| fetch $e | print $'($e) size: ($in | bytes size)' }`. So, I'm not exactly convinced that `par-each` losing the ordering is necessarily essential to it.

At the least, `par-each` could simply only stream a particular result once all results earlier in the ordering are completed. This would at worst be the same as collecting, but at best would have no perf loss.