@rgwood 2023-01-10
> `par-each` could simply only stream a particular result once all results earlier in the ordering are completed. This would at worst be the same as collecting, but at best would have no perf loss.

I think that would be a reasonable default.