@rgwood 2023-01-25
We discussed this in today's team meeting and decided to close it for now, sorry; we don't want to guarantee ordering.

I can see both sides of the issue; there are situations where it makes sense to order the output and situations where it does not. Parallelism is definitely an area that needs more work+attention, and we welcome future contributions in this area.